requirements.txt
requests
python-dotenv
aiogram==3.7.0
apscheduler
BeautifulSoup4